refactor: move parser options to core
