+++ packages/generator/src/index.ts
@@ -1,2 +1,1 @@
 export * from './generator'
-export type { ParserOptions } from './parser-options'
