+++ packages/parser/src/parser.ts
@@ -1,7 +1,7 @@
 import { resolveTsPathPattern } from '@pandacss/config/ts-path'
-import type { ImportResult } from '@pandacss/core'
+import type { ImportResult, ParserOptions } from '@pandacss/core'
 import { BoxNodeMap, box, extract, unbox, type EvaluateOptions, type Unboxed } from '@pandacss/extractor'
-import type { Generator, ParserOptions } from '@pandacss/generator'
+import type { Generator } from '@pandacss/generator'
 import { logger } from '@pandacss/logger'
 import { astish } from '@pandacss/shared'
 import type { ResultItem } from '@pandacss/types'
